+++ packages/terra-profile-image/CHANGELOG.md
@@ -6,6 +6,12 @@ Unreleased
 ### Updated
 * updated jest snapshots
 
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
+
+### Added
+* Added documentation on required peerDependencies
+
 3.12.0  - (July 11, 2019)
 ------------------
 ### Changed