+++ packages/terra-profile-image/CHANGELOG.md
@@ -3,6 +3,21 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
+
+### Added
+* Added documentation on required peerDependencies
+
+3.12.0  - (July 11, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.11.0 - (June 28, 2019)
+------------------
+### Changed
+* Minor dependency version bump
 
 3.10.0 - (June 11, 2019)
 ------------------